refactor(domains): dissolve core/utils + seed tools to plugin + messaging flatten (v0.99.162)
- atomic_io->memory/atomic_write, env_io+project_detect->config,
  redaction->observability, similarity->seed_generation (dominant-caller
  census; 2 blueprint guesses corrected)
- core/tools seed trio -> plugins/seed_generation/tools/ (+mirror tests)
- core/integrations/messaging -> core/messaging (contract TOML updated)
- 46+ importers migrated
