+++ core/cli/tool_handlers/delegated.py
@@ -25,18 +25,21 @@ _DELEGATED_TOOLS: dict[str, tuple[str, str]] = {
     # CSP-2 (2026-05-22) — literature research surface
     "arxiv_search": ("core.tools.arxiv", "ArxivSearchTool"),
     "paper_fetch_arxiv": ("core.tools.arxiv", "ArxivFetchTool"),
-    "geode_seed_pool_search": ("core.tools.seed_pool_search", "SeedPoolSearchTool"),
+    "geode_seed_pool_search": (
+        "plugins.seed_generation.tools.seed_pool_search",
+        "SeedPoolSearchTool",
+    ),
     # CSP-13 (2026-05-23) — Loop 2 (debate-turn) of the seed-generation
     # 3-loop port. Records one debate turn to a per-candidate sidecar
     # JSONL and signals continue/synthesize so the sub-agent's
     # AgenticLoop completes the N-turn budget before writing the seed.
-    "seed_debate_turn": ("core.tools.seed_debate", "SeedDebateTurnTool"),
+    "seed_debate_turn": ("plugins.seed_generation.tools.seed_debate", "SeedDebateTurnTool"),
     # CSP-14 (2026-05-23) — Loop 3 (paper-analysis) of the seed-generation
     # 3-loop port. Freezes one fetched arXiv paper into a git-tracked
     # snapshot under docs/self-improving/petri-bundle/literature/. Cache-hit on
     # content_hash short-circuits re-writes.
     "freeze_paper_snapshot": (
-        "core.tools.literature_snapshot",
+        "plugins.seed_generation.tools.literature_snapshot",
         "FreezePaperSnapshotTool",
     ),
     # profile
