+++ core/config/project_detect.py
