--- core/integrations/__init__.py
@@ -1,5 +1,0 @@
-"""External-system integrations — capability-pure modules wired by the server layer.
-
-Subpackages own one external system each (``messaging`` covers Slack /
-Discord / Telegram via the unified Gateway / Channel binding).
-"""
