+++ core/memory/context.py
@@ -290,7 +290,7 @@ class ContextAssembler:
         if not self._project_root:
             return
         try:
-            from core.utils.project_detect import (
+            from core.config.project_detect import (
                 detect_project_type,
                 get_harness_summary,
             )
