+++ core/memory/project.py
@@ -22,8 +22,8 @@ from datetime import datetime
 from pathlib import Path
 from typing import Any
 
+from core.memory.atomic_write import atomic_write_text
 from core.skills._frontmatter import _FRONTMATTER_RE
-from core.utils.atomic_io import atomic_write_text
 
 log = logging.getLogger(__name__)
