+++ core/memory/user_profile.py
@@ -23,9 +23,9 @@ from datetime import datetime
 from pathlib import Path
 from typing import Any, cast
 
+from core.memory.atomic_write import atomic_write_text
 from core.paths import GLOBAL_USER_PROFILE_DIR
 from core.skills._frontmatter import _FRONTMATTER_RE
-from core.utils.atomic_io import atomic_write_text
 
 log = logging.getLogger(__name__)
