+++ core/memory/vault.py
@@ -24,7 +24,7 @@ from datetime import date
 from pathlib import Path
 from typing import Any, cast
 
-from core.utils.atomic_io import atomic_write_text
+from core.memory.atomic_write import atomic_write_text
 
 log = logging.getLogger(__name__)
