+++ core/messaging/binding.py
@@ -12,8 +12,8 @@ from collections.abc import Awaitable, Callable
 from inspect import isawaitable
 from typing import Any
 
-from core.integrations.messaging.models import ChannelBinding, InboundMessage
 from core.memory.session_key import build_gateway_session_key
+from core.messaging.models import ChannelBinding, InboundMessage
 from core.server.supervised.poller_base import BasePoller
 
 MessageProcessor = Callable[[str, dict[str, Any]], Awaitable[str] | str]
