+++ core/messaging/slack_formatter.py
