+++ core/server/supervised/slack_poller.py
@@ -6,15 +6,15 @@ import logging
 import time
 from typing import TYPE_CHECKING, Any
 
-from core.integrations.messaging.models import InboundMessage
+from core.messaging.models import InboundMessage
 from core.server.supervised.poller_base import BasePoller
 
 log = logging.getLogger(__name__)
 
 if TYPE_CHECKING:
-    from core.integrations.messaging.binding import ChannelManager
     from core.mcp.manager import MCPServerManager
     from core.mcp.notification_port import NotificationPort
+    from core.messaging.binding import ChannelManager
 
 
 class SlackPoller(BasePoller):
@@ -208,7 +208,7 @@ class SlackPoller(BasePoller):
         if self._mcp is None:
             return
         try:
-            from core.integrations.messaging.slack_formatter import markdown_to_slack_mrkdwn
+            from core.messaging.slack_formatter import markdown_to_slack_mrkdwn
 
             text = markdown_to_slack_mrkdwn(text)
             args: dict[str, Any] = {"channel_id": channel_id, "text": text}
