+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "geode-agent"
-version = "0.99.161"
+version = "0.99.162"
 description = "GEODE — 범용 자율 실행 에이전트"
 license = "Apache-2.0"
 readme = "README.md"
@@ -229,7 +229,7 @@ root_package = "core"
 name = "CLI must not import server/messaging (process boundary)"
 type = "forbidden"
 source_modules = ["core.cli"]
-forbidden_modules = ["core.server", "core.integrations.messaging"]
+forbidden_modules = ["core.server", "core.messaging"]
 # v0.52 phase 6 — geode serve 명령이 cli/__init__.py 안에 있어서
 # daemon entry 가 cli/ 의 일부로 남아있음. v0.53 에서 server/entrypoint.py
 # 로 분리하면 이들 ignore 제거.
@@ -237,9 +237,9 @@ ignore_imports = [
     "core.cli.typer_serve -> core.server.supervised.services",
     "core.cli.typer_serve -> core.server.supervised.webhook_handler",
     "core.cli.typer_serve -> core.server.ipc_server.poller",
-    "core.cli.typer_serve -> core.integrations.messaging.binding",
+    "core.cli.typer_serve -> core.messaging.binding",
     "core.cli.scheduler_drain -> core.server.supervised.services",
-    "core.wiring.adapters -> core.integrations.messaging.binding",
+    "core.wiring.adapters -> core.messaging.binding",
 ]
 
 [[tool.importlinter.contracts]]
@@ -294,13 +294,13 @@ ignore_imports = [
 [[tool.importlinter.contracts]]
 name = "Messaging integration is capability-pure (no process imports)"
 type = "forbidden"
-source_modules = ["core.integrations.messaging"]
+source_modules = ["core.messaging"]
 forbidden_modules = ["core.cli", "core.server", "core.agent"]
 # v0.52 phase 6 — channel_manager (구) 의 BasePoller 의존은 사실 channels 가
 # poller 추상화를 알아야 함. v0.53 에서 poller_base 를 channels/ 로 승격할지
 # server/ 에 둘지 결정.
 ignore_imports = [
-    "core.integrations.messaging.binding -> core.server.supervised.poller_base",
+    "core.messaging.binding -> core.server.supervised.poller_base",
 ]
 
 [tool.ruff]
