+++ tests/core/tools/test_redaction.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from core.utils.redaction import redact_secrets
+from core.observability.redaction import redact_secrets
 
 
 class TestRedactSecrets:
