+++ tests/integration/test_no_daemon_print.py
@@ -23,7 +23,7 @@ from pathlib import Path
 DAEMON_DIRS = [
     "core/agent",
     "core/auth",
-    "core/integrations/messaging",  # phase 4 — gateway 분해 결과
+    "core/messaging",  # phase 4 — gateway 분해 결과
     "core/hooks",
     "core/lifecycle",
     "core/llm",
